endpoint: move headerfile-syncing logic to `bpf.go`
Signed-off by: Ian Vernon <ian@cilium.io>
